Use a single resource for managing bulk syncs
Prior to this commit a bulk sync was a separate resource from each of
the schemas (tables) involved. This commit changes that so that the
bulk sync resource encapsulates all of the data needed for the sync.

At this time the source and destination configuration JSON must be
complete: any missing fields will cause Terraform to fail. This will
be addressed in a future commit.